refactor(ui): move locale into i18n family
