+++ npm/ui/core/src/families/i18n/locale/locale-text.ts
