+++ npm/ui/core/src/families/i18n/locale/locale.test.ts
@@ -3,7 +3,7 @@ import assert from "node:assert/strict";
 import { test } from "vite-plus/test";
 import { defineComponent, h, nextTick, ref } from "vue";
 
-import { mountInteraction } from "./testing/mount.ts";
+import { mountInteraction } from "../../../testing/mount.ts";
 import {
   localeTextMatches,
   normalizeLocaleText,
