+++ npm/ui/core/src/family-catalog-overlays.ts
@@ -68,35 +68,6 @@ export const overlayFamilyCatalog = [
     owner: catalogOwner,
   },
   {
-    canonicalName: "locale",
-    title: "Locale and Direction",
-    packageSubpath: "./locale",
-    entryFile: "src/locale.ts",
-    sourceFiles: [
-      "src/locale-provider.vue",
-      "src/locale.ts",
-      "src/locale-runtime.ts",
-      "src/locale-text.ts",
-    ],
-    behaviorContract: "src/locale.behavior.md",
-    tests: ["src/locale.test.ts", "src/locale-ssr.test.ts"],
-    typeTests: ["src/locale.types.test-d.ts"],
-    rendererFixture: "locale-provider.vue",
-    qualityGates: componentQualityGates,
-    bundleBudget: {
-      exportName: "LocaleProvider",
-      retainedSignature: "data-vize-ui.+locale",
-      allowedRetainedFamilies: ["context"],
-      maximumJavaScriptGzipBytes: 1_400,
-      maximumCssGzipBytes: 0,
-    },
-    aliases: ["direction", "rtl", "writing mode", "i18n provider"],
-    upstreamCoverage: ["HTML dir", "HTML lang", "Intl.Locale"],
-    dependencies: ["context"],
-    maturity: "stable",
-    owner: catalogOwner,
-  },
-  {
     canonicalName: "motion",
     title: "Motion Tokens",
     packageSubpath: "./motion",
