+++ npm/ui/core/src/sfc-lint.test.ts
@@ -50,6 +50,7 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/families/feedback/callout/callout.vue",
       "src/families/feedback/status-light/status-light.vue",
       "src/families/form/slider/slider.vue",
+      "src/families/i18n/locale/locale-provider.vue",
       "src/families/layout/icon/icon-button.vue",
       "src/families/layout/icon/icon.vue",
       "src/families/layout/surface/surface.vue",
@@ -81,7 +82,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "src/listbox-item.vue",
       "src/listbox.vue",
       "src/live-region.vue",
-      "src/locale-provider.vue",
       "src/meter.vue",
       "src/pagination-ellipsis.vue",
       "src/pagination-item.vue",
@@ -158,6 +158,7 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "families/feedback/callout/callout.vue",
       "families/feedback/status-light/status-light.vue",
       "families/form/slider/slider.vue",
+      "families/i18n/locale/locale-provider.vue",
       "families/layout/icon/icon-button.vue",
       "families/layout/icon/icon.vue",
       "families/layout/surface/surface.vue",
@@ -189,7 +190,6 @@ test("discovers every SFC with the opinionated Vize contract", async () => {
       "listbox-item.vue",
       "listbox.vue",
       "live-region.vue",
-      "locale-provider.vue",
       "meter.vue",
       "pagination-ellipsis.vue",
       "pagination-item.vue",
